TOPページのリファクタリング
react-iconsのimpprt修正

tailwindの設定ファイル修正
