+++ tailwind.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  purge: ['./public/**/*.html', './pages/**/*.{js,ts,jsx,tsx}', './component/**/*.{js,ts,jsx,tsx}'],
+  purge: ['./public/**/*.html', './src/pages/**/*.{js,ts,jsx,tsx}', './src/component/**/*.{js,ts,jsx,tsx}'],
   darkMode: false, // or 'media' or 'class'
   theme: {
     container: {
